# docs/swagger.yaml
definitions:
...
  dto.CompleteRegistrationRequest:
    properties:
      otp_code:
        type: string
      otp_id:
        type: string
      user_id:
        type: string
    required:
    - otp_code
    - otp_id
    - user_id
    type: object
...
  dto.DetailedAccount:
    properties:
      bonus_money:
        type: integer
      currency:
        type: string
      id:
        type: string
      real_money:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  dto.DetailedUserRegistration:
    properties:
      accounts:
        items:
          $ref: '#/definitions/dto.DetailedAccount'
        type: array
      agent_request_id:
        type: string
      city:
        type: string
      country:
        type: string
      date_of_birth:
        type: string
      default_currency:
        type: string
      email:
        type: string
      first_name:
        type: string
      kyc_status:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      postal_code:
        type: string
      profile_picture:
        type: string
      referal_type:
        type: string
      refered_by_code:
        type: string
      referral_code:
        type: string
      state:
        type: string
      street_address:
        type: string
      type:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  dto.RegistrationCompleteResponse:
    properties:
      access_token:
        type: string
      is_new_user:
        type: boolean
      message:
        type: string
      refresh_token:
        type: string
      user_id:
        type: string
      user_profile:
        $ref: '#/definitions/dto.UserProfile'
    type: object
  dto.RegistrationPendingResponse:
    properties:
      email:
        type: string
      expires_at:
        type: string
      message:
        type: string
      otp_id:
        type: string
      resend_after:
        type: string
      user_id:
        type: string
    type: object
...
  dto.ResendOTPRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.ResendOTPResponse:
    properties:
      email:
        type: string
      expires_at:
        type: string
      message:
        type: string
      otp_id:
        type: string
      resend_after:
        type: string
    type: object
...
  dto.Type:
    enum:
    - PLAYER
    - OTG_AGENT
    - INFLUENCER
    type: string
    x-enum-varnames:
    - PLAYER
    - OTG_AGENT
    - INFLUENCER
...
  dto.UserProfile:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        minLength: 8
        type: string
      profile_picture:
        type: string
      referal_type:
        $ref: '#/definitions/dto.Type'
      refered_by_code:
        type: string
      referral_code:
        type: string
      type:
        $ref: '#/definitions/dto.Type'
      user_id:
        type: string
    required:
    - phone_number
    type: object
...
info:
  contact: {}
paths:
...
  /register:
    post:
      consumes:
      - application/json
      description: Start user registration process and send email verification OTP
      parameters:
      - description: User registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DetailedUserRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RegistrationPendingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Initiate user registration
      tags:
      - User
  /register/complete:
    post:
      consumes:
      - application/json
      description: Verify OTP and complete user registration process
      parameters:
      - description: Complete registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CompleteRegistrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RegistrationCompleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Complete user registration
      tags:
      - User
  /register/resend-verification:
    post:
      consumes:
      - application/json
      description: Resend verification email for pending user registration
      parameters:
      - description: Resend verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResendOTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Resend verification email
      tags:
      - User
...
swagger: "2.0"